Make `but` argument handling available from its library.
