+++ crates/but/Cargo.toml
@@ -9,6 +9,7 @@ authors = ["GitButler <gitbutler@gitbutler.com>"]
 readme = "../../README.md"
 publish = false
 rust-version = "1.89"
+
 [[bin]]
 name = "but"
 path = "src/main.rs"
