+++ crates/gitbutler-tauri/src/main.rs
@@ -96,14 +96,6 @@ fn main() {
                         tracing::info!("SHELL env: {var:?}", var = std::env::var_os("SHELL"));
                     }
 
-                    // On MacOS, in dev mode with debug assertions, we encounter popups each time
-                    // the binary is rebuilt. To counter that, use a git-credential based implementation.
-                    // This isn't an issue for actual release build (i.e. nightly, production),
-                    // hence the specific condition.
-                    if cfg!(debug_assertions) && cfg!(target_os = "macos") {
-                        but_secret::secret::git_credentials::setup().ok();
-                    }
-
                     // SAFETY(qix-): This is safe because we're initializing the askpass broker here,
                     // SAFETY(qix-): before any other threads would ever access it.
                     unsafe {
